Message Passing/modules/location_microservice/requirements.txt
kafka-python==2.0.2
marshmallow==3.15.0
psycopg2-binary==2.9.3
